# Performance improvements for framework