--- a/Common/Securities/Positions/CompositePositionGroupResolver.cs
+++ b/Common/Securities/Positions/CompositePositionGroupResolver.cs
@@ -116,7 +116,7 @@ namespace QuantConnect.Securities.Positions
         /// <param name="currentPositions">The currently grouped positions</param>
         /// <param name="group">The grouped positions when this resolver is able to, otherwise null</param>
         /// <returns>True if this resolver can group the specified positions, otherwise false</returns>
-        public bool TryGroup(IReadOnlyCollection<IPosition> newPositions, PositionGroupCollection currentPositions, out IPositionGroup group)
+        public bool TryGroup(IReadOnlyList<IPosition> newPositions, PositionGroupCollection currentPositions, out IPositionGroup group)
         {
             foreach (var resolver in _resolvers)
             {