--- a/Common/Securities/Positions/IPositionGroupResolver.cs
+++ b/Common/Securities/Positions/IPositionGroupResolver.cs
@@ -31,7 +31,7 @@ namespace QuantConnect.Securities.Positions
         /// <param name="currentPositions">The currently grouped positions</param>
         /// <param name="group">The grouped positions when this resolver is able to, otherwise null</param>
         /// <returns>True if this resolver can group the specified positions, otherwise false</returns>
-        bool TryGroup(IReadOnlyCollection<IPosition> newPositions, PositionGroupCollection currentPositions, out IPositionGroup group);
+        bool TryGroup(IReadOnlyList<IPosition> newPositions, PositionGroupCollection currentPositions, out IPositionGroup group);
 
         /// <summary>
         /// Resolves the position groups that exist within the specified collection of positions.