--- a/Common/Securities/Positions/OptionStrategyPositionGroupResolver.cs
+++ b/Common/Securities/Positions/OptionStrategyPositionGroupResolver.cs
@@ -57,7 +57,7 @@ namespace QuantConnect.Securities.Positions
         /// <param name="currentPositions">The currently grouped positions</param>
         /// <param name="group">The grouped positions when this resolver is able to, otherwise null</param>
         /// <returns>True if this resolver can group the specified positions, otherwise false</returns>
-        public bool TryGroup(IReadOnlyCollection<IPosition> newPositions, PositionGroupCollection currentPositions, out IPositionGroup @group)
+        public bool TryGroup(IReadOnlyList<IPosition> newPositions, PositionGroupCollection currentPositions, out IPositionGroup @group)
         {
             var impactedGroups = GetImpactedGroups(currentPositions, newPositions);
             var positionsToConsiderInNewGroup = impactedGroups.SelectMany(positionGroup => positionGroup.Positions);